+++ plugins/clickhouse/pkg/instance/logs.go
@@ -3,11 +3,20 @@ package instance
 import (
 	"fmt"
 	"strings"
+
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
 )
 
 var (
 	defaultFields  = []string{"timestamp", "cluster", "namespace", "app", "pod_name", "container_name", "host", "log"}
 	defaultColumns = "timestamp, cluster, namespace, app, pod_name, container_name, host, fields_string.key, fields_string.value, fields_number.key, fields_number.value, log"
+
+	fieldsMetric = promauto.NewCounterVec(prometheus.CounterOpts{
+		Namespace: "kobs",
+		Name:      "clickhouse_fields_total",
+		Help:      "Number how often a field was used in a query.",
+	}, []string{"field"})
 )
 
 // parseLogsQuery parses the given query string and return the conditions for the where statement in the sql query. We
@@ -122,6 +131,12 @@ func splitOperator(condition string) (string, error) {
 func handleConditionParts(key, value, operator string) (string, error) {
 	key = strings.TrimSpace(key)
 	value = strings.TrimSpace(value)
+
+	// The kobs_clickhouse_fields_total metric can be used to determine how often a field is used. This information can
+	// then be used to create an additional column for this field via the following SQL commands:
+	// ALTER TABLE logs.logs ON CLUSTER '{cluster}' ADD COLUMN <FIELD> String DEFAULT fields_string.value[indexOf(fields_string.key, '<FIELD>')];
+	// ALTER TABLE logs.logs ON CLUSTER '{cluster}' ADD COLUMN <FIELD> String DEFAULT fields_number.value[indexOf(fields_number.key, '<FIELD>')];
+	fieldsMetric.WithLabelValues(key).Inc()
 
 	if contains(defaultFields, key) {
 		if operator == "=~" {